Refactor(#53) : 퍼스널 페이지 뮤테이션 분리
